pyproject.toml
[tool.poetry]
name = "aqi-notifier"
version = "0.1.0"
description = ""
authors = ["Alla Barbalat <allabarbalat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
rich = "^13.7.1"
pytest = "^8.2.0"
vcrpy = "^6.0.1"
python-aqi = "^0.6.1"
telnyx = "^2.1.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"

[tool.mypy]
[[tool.mypy.overrides]]
module = [
    "vcr",
    "requests",
    "aqi",
    "telnyx"
]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
